feat: add session restoration and connectivity handling for improved authentication flow
